+++ tests-e2e/src/node/api_clients/rest.py
@@ -54,7 +54,7 @@ class REST(BaseClient):
         return self.rest_call("post", "relay/v1/auto/messages", json.dumps(message))
 
     def send_light_push_message(self, payload):
-        return self.rest_call("post", "lightpush/v1/message", json.dumps(payload))
+        return self.rest_call("post", "lightpush/v3/message", json.dumps(payload))
 
     def get_relay_messages(self, pubsub_topic):
         get_messages_response = self.rest_call("get", f"relay/v1/messages/{quote(pubsub_topic, safe='')}")