+++ tests/app/test_all.nim
@@ -1,3 +1,3 @@
 {.used.}
 
-import ./test_app, ./test_validators, ./test_cli_args
+import ./test_app, ./test_validators, ./test_cli_args, ./test_rlnkeystore_conf